fix: few more to 'npm run check'
